Print the pipeline web URL on each status check
In case the pipeline takes a long time or is stuck, and one
opens the GH job view there is no way to see the past logs
without canceling the job, or waiting for it to finish.

To make it easier in such cases to go to the gitlab job page to
see what is going on print the pipeline URL on each status check.